feat(core): switch default config to v2
