+++ e2e/linter/src/linter.test.ts
@@ -77,8 +77,8 @@ describe('Linter', () => {
     // This test is only relevant for the deprecated lint builder,
     // so we need to patch the workspace.json to use it
     const workspaceJson = readJson(`workspace.json`);
-    workspaceJson.projects[myapp].architect.lint = {
-      builder: '@nrwl/linter:lint',
+    workspaceJson.projects[myapp].targets.lint = {
+      executor: '@nrwl/linter:lint',
       options: {
         linter: 'eslint',
         tsConfig: [
