+++ e2e/storybook/src/storybook.test.ts
@@ -3,166 +3,13 @@ import {
   newProject,
   readFile,
   runCLI,
-  supportUi,
   tmpProjPath,
   uniq,
 } from '@nrwl/e2e/utils';
-import { mkdirSync, writeFileSync } from 'fs';
+import { writeFileSync } from 'fs';
 
 describe('Storybook schematics', () => {
-  describe('running Storybook and Cypress', () => {
-    it('should execute e2e tests using Cypress running against Storybook', () => {
-      newProject();
-
-      const myapp = uniq('myapp');
-      runCLI(`generate @nrwl/angular:app ${myapp} --no-interactive`);
-
-      const mylib = uniq('test-ui-lib');
-      createTestUILib(mylib);
-      const mylib2 = uniq('test-ui-lib-react');
-      runCLI(`generate @nrwl/react:lib ${mylib2} --no-interactive`);
-      runCLI(
-        `generate @nrwl/react:component Button --project=${mylib2} --no-interactive`
-      );
-      writeFileSync(
-        tmpProjPath(`libs/${mylib2}/src/lib/button.tsx`),
-        `
-          import React from 'react';
-
-            import './button.css';
-            
-            export type ButtonStyle = 'default' | 'primary' | 'warning';
-            
-            /* eslint-disable-next-line */
-            export interface ButtonProps {
-              text?: string;
-              style?: ButtonStyle;
-              padding?: number;
-            }
-            
-            export const Button = (props: ButtonProps) => {
-              return (
-                <button className={props.style} style={{ padding: \`\${props.padding}px\` }}>
-                  {props.text}
-                </button>
-              );
-            };
-            
-            export default Button;            
-            `
-      );
-      writeFileSync(
-        tmpProjPath(`libs/${mylib2}/src/lib/button.stories.tsx`),
-        `
-            import React from 'react';
-            import { Button, ButtonStyle } from './button';
-            import { text, number } from '@storybook/addon-knobs';
-            
-            export default { title: 'Button' };
-            
-            export const primary = () => (
-              <Button
-                padding={number('Padding', 0)}
-                style={text('Style', 'default') as ButtonStyle}
-                text={text('Text', 'Click me')}
-                // padding='0'
-                // style='default'
-                // text='Click me'
-              />
-            );
-            `
-      );
-
-      runCLI(
-        `generate @nrwl/angular:storybook-configuration ${mylib} --configureCypress --generateStories --generateCypressSpecs --no-interactive`
-      );
-      runCLI(
-        `generate @nrwl/angular:stories ${mylib} --generateCypressSpecs --no-interactive`
-      );
-
-      writeFileSync(
-        tmpProjPath(
-          `apps/${mylib}-e2e/src/integration/test-button/test-button.component.spec.ts`
-        ),
-        `
-            describe('test-ui-lib3726865', () => {
-
-          it('should render the component', () => {
-            cy.visit('/iframe.html?id=testbuttoncomponent--primary&knob-buttonType=button&knob-style=default&knob-age&knob-isDisabled=false');
-            cy.get('proj-test-button').should('exist');
-            cy.get('button').should('not.be.disabled');
-            cy.get('button').should('have.class', 'default');
-            cy.contains('You are 0 years old.');
-          });
-          it('should adjust the knobs', () => {
-            cy.visit('/iframe.html?id=testbuttoncomponent--primary&knob-buttonType=button&knob-style=primary&knob-age=10&knob-isDisabled=true');
-            cy.get('button').should('be.disabled');
-            cy.get('button').should('have.class', 'primary');
-            cy.contains('You are 10 years old.');
-          });
-        });
-        `
-      );
-
-      runCLI(
-        `generate @nrwl/react:storybook-configuration ${mylib2} --configureCypress --no-interactive`
-      );
-
-      mkdirSync(tmpProjPath(`apps/${mylib2}-e2e/src/integration`));
-      writeFileSync(
-        tmpProjPath(`apps/${mylib2}-e2e/src/integration/button.spec.ts`),
-        `
-        describe('react-ui', () => {
-          it('should render the component', () => {
-            cy.visit(
-              '/iframe.html?id=button--primary&knob-Style=default&knob-Padding&knob-Text=Click%20me'
-            );
-            cy.get('button').should('exist');
-            cy.get('button').should('have.class', 'default');
-          });
-          it('should adjust the knobs', () => {
-            cy.visit(
-              '/iframe.html?id=button--primary&knob-Style=primary&knob-Padding=10&knob-Text=Other'
-            );
-            cy.get('button').should('have.class', 'primary');
-          });
-        });
-        `
-      );
-
-      if (supportUi()) {
-        expect(runCLI(`run ${mylib}-e2e:e2e --no-watch`)).toContain(
-          'All specs passed!'
-        );
-      }
-
-      runCLI(`run ${mylib}:build-storybook`);
-
-      checkFilesExist(`dist/storybook/${mylib}/index.html`);
-      expect(readFile(`dist/storybook/${mylib}/index.html`)).toContain(
-        `<title>Storybook</title>`
-      );
-    }, 1000000);
-  });
-
   describe('build storybook', () => {
-    it('should build an Angular based storybook', () => {
-      newProject();
-
-      const angularStorybookLib = uniq('test-ui-lib');
-      createTestUILib(angularStorybookLib);
-      runCLI(
-        `generate @nrwl/angular:storybook-configuration ${angularStorybookLib} --generateStories --no-interactive`
-      );
-
-      // build Angular lib
-      runCLI(`run ${angularStorybookLib}:build-storybook`);
-      checkFilesExist(`dist/storybook/${angularStorybookLib}/index.html`);
-      expect(
-        readFile(`dist/storybook/${angularStorybookLib}/index.html`)
-      ).toContain(`<title>Storybook</title>`);
-    }, 1000000);
-
     it('should build a React based storybook', () => {
       newProject();
 
@@ -248,62 +95,6 @@ describe('Storybook schematics', () => {
       checkFilesExist(`dist/storybook/${reactStorybookLib}/index.html`);
       expect(
         readFile(`dist/storybook/${reactStorybookLib}/index.html`)
-      ).toContain(`<title>Storybook</title>`);
-    }, 1000000);
-
-    it('should build an Angular based storybook that references another lib', () => {
-      newProject();
-
-      const angularStorybookLib = uniq('test-ui-lib');
-      createTestUILib(angularStorybookLib);
-      runCLI(
-        `generate @nrwl/angular:storybook-configuration ${angularStorybookLib} --generateStories --no-interactive`
-      );
-
-      // create another lib with a component
-      const anotherTestLib = uniq('test-another-lib');
-      runCLI(`g @nrwl/angular:library ${anotherTestLib} --no-interactive`);
-      runCLI(
-        `g @nrwl/angular:component my-test-cmp --project=${anotherTestLib} --no-interactive`
-      );
-
-      // update index.ts and export it
-      writeFileSync(
-        tmpProjPath(`libs/${anotherTestLib}/src/index.ts`),
-        `
-            export * from './lib/my-test-cmp/my-test-cmp.component';
-        `
-      );
-
-      // create a story in the first lib to reference the cmp from the 2nd lib
-      writeFileSync(
-        tmpProjPath(
-          `libs/${angularStorybookLib}/src/lib/myteststory.stories.ts`
-        ),
-        `
-            import { MyTestCmpComponent } from '@proj/${anotherTestLib}';
-
-            export default {
-              title: 'My Test Cmp',
-            };
-            
-            let x = 'hi';
-            
-            export const primary = () => ({
-              moduleMetadata: {
-                imports: [],
-              },
-              component: MyTestCmpComponent,
-              props: {},
-            });
-        `
-      );
-
-      // build Angular lib
-      runCLI(`run ${angularStorybookLib}:build-storybook`);
-      checkFilesExist(`dist/storybook/${angularStorybookLib}/index.html`);
-      expect(
-        readFile(`dist/storybook/${angularStorybookLib}/index.html`)
       ).toContain(`<title>Storybook</title>`);
     }, 1000000);
   });
