+++ packages/devkit/ngcli-adapter.ts
@@ -1,1 +1,4 @@
-export { wrapAngularDevkitSchematic } from '@nrwl/tao/src/commands/ngcli-adapter';
+export {
+  wrapAngularDevkitSchematic,
+  NxScopedHost,
+} from '@nrwl/tao/src/commands/ngcli-adapter';
