+++ packages/workspace/src/schematics/workspace/files/__workspaceFile__.json
@@ -1,5 +1,10 @@
 {
+  <% if(cli === 'angular') { %>
   "version": 1,
+  <% } %>
+  <% if(cli === 'nx') { %>
+  "version": 2,
+  <% } %>
   "projects": {},
   "cli": {
     "defaultCollection": "@nrwl/workspace"
